demo-compose-2/app/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
Faker==37.0.0
fastapi==0.115.11
h11==0.14.0
idna==3.10
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
sniffio==1.3.1
starlette==0.46.1
typing_extensions==4.12.2
tzdata==2025.1
uvicorn==0.34.0
